Add user roles, mood, and product models; implement migrations and seeders

+++ app/Helpers/helpers.php
@@ -1,8 +1,8 @@
 <?php
 
 if (! function_exists('format_rupiah')) {
-    function format_rupiah(float|int $amount): string
+    function format_rupiah(int|float $amount): string
     {
-        return 'Rp '.number_format($amount, 0, ',', '.');
+        return 'Rp ' . number_format($amount, 0, ',', '.');
     }
 }
